# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a voda api server.",
        "title": "Voice Of Diary API (voda)",
        "contact": {
            "name": "API Support",
            "url": "https://minkj1992.github.io",
            "email": "minkj1992@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "exchange-diary-b4mzhzbzcq-du.a.run.app",
    "basePath": "/v1",
    "paths": {
        "/authentication/login/{auth_type}": {
            "get": {
                "description": "해당 auth_type의 login form을 기입 후, redirect되는 곳으로 기입한 email 그리고 요청한 auth_type에 대한 auth code를 발급한다.\n클라이언트 URL에 query string으로 auth code가 담긴 채로 redirect하여 로그인 및 회원가입이 진행되도록 한다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "kakao | google | apple",
                        "name": "auth_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/authentication/mock": {
            "post": {
                "description": "클라 테스트용. 주어진 email이 db에 없으면 회원가입 프로세스 동시에 진행\nAccessToken을 사용해서 헤더에 {\"Authorization\": AccessToken} 넣어주면 된다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "(debug) mock login / register",
                "parameters": [
                    {
                        "description": "모킹할 유저정보",
                        "name": "member",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.mockMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.mockMemberResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/login/{auth_type}": {
            "get": {
                "description": "요청한 auth_type의 login form URL로 redirect됩니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "LoginForm Redirect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "kakao | google | apple",
                        "name": "auth_type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
        "/rooms/{room_id}/files": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "파일 업로드",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "upload static files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "room_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "1.5",
                        "name": "audioPitch",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "LastDayOnEarth",
                        "name": "audioTitle",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "ad5bb198-942f-4ddf-a248-3aaa4bba3b9b",
                        "name": "audioUUID",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "e4947e0c-490b-4588-a14d-e74dd3b8371f",
                        "name": "photoUUID",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "photo",
                        "name": "photo",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "audio",
                        "name": "audio",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.filePostResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.filePostResponse": {
            "type": "object",
            "properties": {
                "audioDownloadURL": {
                    "type": "string"
                },
                "audioURL": {
                    "type": "string"
                },
                "photoDownloadURL": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                }
            }
        },
...
        "controller.mockMemberRequest": {
            "type": "object",
            "properties": {
                "deviceToken": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controller.mockMemberResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "authCode": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}